requirements.txt
pyfrc
pytest-cov